requirements.txt
djangorestframework
django
psycopg2
phonenumbers
django-phonenumber-field
